+++ R/metabin.R
@@ -109,6 +109,8 @@
 #' @param adhoc.hakn.pi A character string indicating whether an
 #'   \emph{ad hoc} variance correction should be applied for
 #'   prediction interval (see \code{\link{meta-package}}).
+#' @param seed.predict A numeric value used as seed to calculate
+#'   bootstrap prediction interval (see \code{\link{meta-package}}).
 #' @param method.bias A character string indicating which test for
 #'   funnel plot asymmetry is to be used. Either \code{"Begg"},
 #'   \code{"Egger"}, \code{"Thompson"}, \code{"Schwarzer"},
@@ -152,6 +154,9 @@
 #'   results of test for subgroup differences.
 #' @param prediction.subgroup A logical indicating whether prediction
 #'   intervals should be printed for subgroups.
+#' @param seed.predict.subgroup A numeric vector providing seeds to
+#'   calculate bootstrap prediction intervals within subgroups. Must
+#'   be of same length as the number of subgroups.
 #' @param byvar Deprecated argument (replaced by 'subgroup').
 #' @param hakn Deprecated argument (replaced by 'method.random.ci').
 #' @param adhoc.hakn Deprecated argument (replaced by
@@ -642,6 +647,7 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
                     level.predict = gs("level.predict"),
                     method.predict = gs("method.predict"),
                     adhoc.hakn.pi = gs("adhoc.hakn.pi"),
+                    seed.predict = NULL,
                     ##
                     method.bias = ifelse(sm == "OR", "Harbord",
                                   ifelse(sm == "DOR", "Deeks",
@@ -667,6 +673,8 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
                     sep.subgroup = gs("sep.subgroup"),
                     test.subgroup = gs("test.subgroup"),
                     prediction.subgroup = gs("prediction.subgroup"),
+                    seed.predict.subgroup = NULL,
+                    ##
                     byvar, hakn, adhoc.hakn,
                     ##
                     print.CMH = gs("print.CMH"),
@@ -697,6 +705,7 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
   missing.method.tau <- missing(method.tau)
   method.tau <- setchar(method.tau, c(gs("meth4tau"), "KD"))
   ##
+  missing.tau.common <- missing(tau.common)
   tau.common <- replaceNULL(tau.common, FALSE)
   chklogical(tau.common)
   ##
@@ -1374,31 +1383,16 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
     three.level <- TRUE
   ##
   if (three.level) {
+    chkmlm(method.tau, missing.method.tau, method.predict,
+           by, tau.common, missing.tau.common,
+           method, missing.method)
+    ##
     common <- FALSE
+    method <- "Inverse"
+    is.glmm <- FALSE
     ##
-    if (method != "Inverse") {
-      if (!missing.method)
-        warning("Inverse variance method used in three-level model.",
-                call. = FALSE)
-      method <- "Inverse"
-      is.glmm <- FALSE
-    }
-    ##
-    if (!(method.tau %in% c("REML", "ML"))) {
-      if (!missing.method.tau)
-        warning("For three-level model, argument 'method.tau' set to ",
-                "\"REML\".",
-                call. = FALSE)
+    if (!(method.tau %in% c("REML", "ML")))
       method.tau <- "REML"
-    }
-    ##
-    if (by & !tau.common) {
-      if (!missing(tau.common))
-        warning("For three-level model, argument 'tau.common' set to ",
-                "\"TRUE\".",
-                call. = FALSE)
-      tau.common <- TRUE
-    }
   }
   
   
@@ -1625,6 +1619,7 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
                level.predict = level.predict,
                method.predict = method.predict,
                adhoc.hakn.pi = adhoc.hakn.pi,
+               seed.predict = seed.predict,
                ##
                method.bias = method.bias,
                ##
@@ -1772,12 +1767,13 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
     res$tau.common <- tau.common
     ##
     if (!tau.common) {
-      res <- c(res, subgroup(res))
+      res <- c(res, subgroup(res, seed = seed.predict.subgroup))
       if (res$three.level)
         res <- setNA3(res)
     }
     else if (!is.null(tau.preset))
-      res <- c(res, subgroup(res, tau.preset))
+      res <-
+        c(res, subgroup(res, tau.preset, seed = seed.predict.subgroup))
     else {
       if (is.glmm)
         res <- c(res,
@@ -1788,7 +1784,8 @@ metabin <- function(event.e, n.e, event.c, n.c, studlab,
                  subgroup(res, NULL,
                           factor(res$subgroup, bylevs(res$subgroup))))
       else
-        res <- c(res, subgroup(res, hcc$tau.resid))
+        res <-
+          c(res, subgroup(res, hcc$tau.resid, seed = seed.predict.subgroup))
     }
     ##
     if (tau.common && is.null(tau.preset))