+++ R/metagen.R
@@ -79,6 +79,8 @@
 #' @param adhoc.hakn.pi A character string indicating whether an
 #'   \emph{ad hoc} variance correction should be applied for
 #'   prediction interval (see \code{\link{meta-package}}).
+#' @param seed.predict A numeric value used as seed to calculate
+#'   bootstrap prediction interval (see \code{\link{meta-package}}).
 #' @param null.effect A numeric value specifying the effect under the
 #'   null hypothesis.
 #' @param method.bias A character string indicating which test is to
@@ -118,8 +120,8 @@
 #'   (see Details).
 #' @param transf A logical indicating whether inputs for arguments
 #'   \code{TE}, \code{lower} and \code{upper} are already
-#'   appropriately transformed to conduct the meta-analysis or on
-#'   the original scale. If \code{transf = TRUE} (default), inputs are
+#'   appropriately transformed to conduct the meta-analysis or on the
+#'   original scale. If \code{transf = TRUE} (default), inputs are
 #'   expected to be log odds ratios instead of odds ratios for
 #'   \code{sm = "OR"} and Fisher's z transformed correlations instead
 #'   of correlations for \code{sm = "ZCOR"}, for example.
@@ -169,6 +171,9 @@
 #'   results of test for subgroup differences.
 #' @param prediction.subgroup A logical indicating whether prediction
 #'   intervals should be printed for subgroups.
+#' @param seed.predict.subgroup A numeric vector providing seeds to
+#'   calculate bootstrap prediction intervals within subgroups. Must
+#'   be of same length as the number of subgroups.
 #' @param byvar Deprecated argument (replaced by 'subgroup').
 #' @param id Deprecated argument (replaced by 'cluster').
 #' @param adhoc.hakn Deprecated argument (replaced by
@@ -624,6 +629,7 @@ metagen <- function(TE, seTE, studlab,
                     level.predict = gs("level.predict"),
                     method.predict = gs("method.predict"),
                     adhoc.hakn.pi = gs("adhoc.hakn.pi"),
+                    seed.predict = NULL,
                     ##
                     null.effect = 0,
                     ##
@@ -660,6 +666,8 @@ metagen <- function(TE, seTE, studlab,
                     sep.subgroup = gs("sep.subgroup"),
                     test.subgroup = gs("test.subgroup"),
                     prediction.subgroup = gs("prediction.subgroup"),
+                    seed.predict.subgroup = NULL,
+                    ##
                     byvar, id, adhoc.hakn,
                     ##
                     keepdata = gs("keepdata"),
@@ -703,6 +711,7 @@ metagen <- function(TE, seTE, studlab,
   missing.method.tau <- missing(method.tau)
   method.tau <- setchar(method.tau, gs("meth4tau"))
   ##
+  missing.tau.common <- missing(tau.common)
   tau.common <- replaceNULL(tau.common, FALSE)
   chklogical(tau.common)
   ##
@@ -723,6 +732,9 @@ metagen <- function(TE, seTE, studlab,
     is.installed.package("pimeta", argument = "method.predict", value = "NNF")
   ##
   adhoc.hakn.pi <- setchar(adhoc.hakn.pi, gs("adhoc4hakn.pi"))
+  ##
+  if (!is.null(seed.predict))
+    chknumeric(seed.predict, length = 1)
   ##
   chknumeric(null.effect, length = 1)
   ##
@@ -1567,13 +1579,13 @@ metagen <- function(TE, seTE, studlab,
   }
   ##
   if (three.level) {
-    if (!(method.tau %in% c("REML", "ML"))) {
-      if (!missing.method.tau)
-        warning("For three-level model, argument 'method.tau' set to ",
-                "\"REML\".",
-                call. = FALSE)
+    chkmlm(method.tau, missing.method.tau, method.predict,
+           by, tau.common, missing.tau.common)
+    ##
+    common <- FALSE
+    ##
+    if (!(method.tau %in% c("REML", "ML")))
       method.tau <- "REML"
-    }
   }
   ##
   if (by) {
@@ -1920,7 +1932,8 @@ metagen <- function(TE, seTE, studlab,
         else if (method.predict[i] == "NNF") {
           res.pima <- pimeta::pima(TE[!exclude], seTE[!exclude],
                                    method = "boot",
-                                   alpha = 1 - level.predict)
+                                   alpha = 1 - level.predict,
+                                   seed = seed.predict)
           ##
           pi.i <- as.data.frame(ci(1, NA, level = level.predict))
           pi.i$seTE <-NA
@@ -2238,6 +2251,8 @@ metagen <- function(TE, seTE, studlab,
               approx.TE = approx.TE,
               approx.seTE = approx.seTE,
               ##
+              seed.predict = seed.predict,
+              ##
               warn = warn,
               call = match.call(),
               version = packageDescription("meta")$Version,
@@ -2266,19 +2281,21 @@ metagen <- function(TE, seTE, studlab,
     res$prediction.subgroup <- prediction.subgroup
     ##
     if (!tau.common) {
-      res <- c(res, subgroup(res))
+      res <- c(res, subgroup(res, seed = seed.predict.subgroup))
       if (res$three.level)
         res <- setNA3(res)
     }
     else if (!is.null(tau.preset))
-      res <- c(res, subgroup(res, tau.preset))
+      res <-
+        c(res, subgroup(res, tau.preset, seed = seed.predict.subgroup))
     else {
       if (three.level)
         res <- c(res,
                  subgroup(res, NULL,
                           factor(res$subgroup, bylevs(res$subgroup))))
       else
-        res <- c(res, subgroup(res, hcc$tau.resid))
+        res <-
+          c(res, subgroup(res, hcc$tau.resid, seed = seed.predict.subgroup))
     }
     ##
     if (tau.common && is.null(tau.preset))